[feat]: Add useSelectBox hook
