+++ src/hooks/useSelectBox.tsx
@@ -1,0 +1,18 @@
+import { Dispatch, SetStateAction } from 'react';
+import { useCallback } from 'react';
+
+const useSelectBox = (setIsOpen: Dispatch<SetStateAction<boolean>>) => {
+  const handleClcikWithoutSelectBox = useCallback((e: Event) => {
+    const target = e.target as HTMLElement;
+
+    if (target.className && !target.className.includes('custom-selected-box')) {
+      setIsOpen(false);
+    }
+  }, []);
+
+  return {
+    handleClcikWithoutSelectBox,
+  };
+};
+
+export default useSelectBox;
